reorganize gotpl funcs by type
